Extracted json-sources.js creation
